backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic
firebase-admin
reportlab
azure-storage-blob
azure-identity
azure-communication-email